requirements.txt
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
python-frontmatter==1.1.0
PyYAML==6.0.1
requests==2.31.0
smmap==5.0.1
urllib3==2.2.1
